+++ packages/remix-dev/compiler/css/compiler.ts
@@ -1,6 +1,5 @@
 import { builtinModules as nodeBuiltins } from "module";
 import * as esbuild from "esbuild";
-import { nodeModulesPolyfillPlugin } from "esbuild-plugins-node-modules-polyfill";
 
 import type { RemixConfig } from "../../config";
 import { getAppDependencies } from "../../dependencies";
@@ -19,6 +18,7 @@ import {
 } from "./plugins/bundleEntry";
 import type { Context } from "../context";
 import { isBundle } from "./bundle";
+import { writeMetafile } from "../analysis";
 
 const getExternals = (config: RemixConfig): string[] => {
   // For the browser build, exclude node built-ins that don't have a
@@ -80,8 +80,11 @@ const createEsbuildConfig = (ctx: Context): esbuild.BuildOptions => {
       absoluteCssUrlsPlugin(),
       externalPlugin(/^https?:\/\//, { sideEffects: false }),
       mdxPlugin(ctx),
+      // Skip compilation of common packages/scopes known not to include CSS imports
+      emptyModulesPlugin(ctx, /^(@remix-run|react|react-dom)(\/.*)?$/, {
+        includeNodeModules: true,
+      }),
       emptyModulesPlugin(ctx, /\.server(\.[jt]sx?)?$/),
-      nodeModulesPolyfillPlugin(),
       externalPlugin(/^node:.*/, { sideEffects: false }),
     ],
     supported: {
@@ -94,9 +97,11 @@ export let create = async (ctx: Context) => {
   let compiler = await esbuild.context({
     ...createEsbuildConfig(ctx),
     write: false,
+    metafile: true,
   });
   let compile = async () => {
-    let { outputFiles } = await compiler.rebuild();
+    let { outputFiles, metafile } = await compiler.rebuild();
+    writeMetafile(ctx, "metafile.css.json", metafile);
     let bundleOutputFile = outputFiles.find((outputFile) =>
       isBundle(ctx, outputFile, ".css")
     );