+++ packages/remix-css-bundle/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @remix-run/css-bundle
 
+## 1.18.1
+
+## 1.18.0
+
+## 1.17.1
+
+### Patch Changes
+
+- Lazily generate CSS bundle when import of `@remix-run/css-bundle` is detected ([#6535](https://github.com/remix-run/remix/pull/6535))
+
 ## 1.17.0
 
 ### Patch Changes