+++ packages/remix-eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # `@remix-run/eslint-config`
 
+## 1.18.1
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.18.1) for an overview of all changes in v1.18.1.
+
+## 1.18.0
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.18.0) for an overview of all changes in v1.18.0.
+
+## 1.17.1
+
+No significant changes to this package were made in this release. [See the releases page on GitHub](https://github.com/remix-run/remix/releases/tag/remix%401.17.1) for an overview of all changes in v1.17.1.
+
 ## 1.17.0
 
 ### Patch Changes